+++ src/index.ts
@@ -1,3 +1,5 @@
 export * from "./core/api";
 export * from "./core/quickLinkers";
 export * from "./core/defs";
+export { default as intersectBody } from "./core/graphql/intersectDeep";
+export { mergeDeep as mergeBody } from "./core/graphql/mergeDeep";